+++ CHANGES.rst
@@ -1,6 +1,9 @@
 1.2.0 (unreleased)
 ==================
 
+- New ``stsynphot.constant`` module to store pre-defined collecting
+  area values for HST and JWST. [#153]
+
 - Dropped support for Python 3.6 and 3.7. Minimum supported Python
   version is now 3.8. [#161]